requirements.txt
# =============================================================================
# Multi-Agent Financial Research Assistant — Dependencies
# =============================================================================

# --- Core LLM & Agent Framework ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# --- Data Sources ---
httpx>=0.27.0
yfinance>=0.2.40

# --- Vector Store ---
faiss-cpu>=1.8.0

# --- NLP & Embeddings ---
tiktoken>=0.7.0

# --- Data Processing ---
pandas>=2.2.0
numpy>=1.26.0

# --- Web UI ---
streamlit>=1.38.0

# --- API Server ---
fastapi>=0.115.0
uvicorn>=0.30.0

# --- Observability ---
langsmith>=0.1.0

# --- Configuration ---
python-dotenv>=1.0.0

# --- Testing ---
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# --- Dev Tools ---
ruff>=0.6.0
mypy>=1.11.0
